stage

+++ apps/web/src/hooks/useCurrentBlockTimestamp.ts
@@ -4,5 +4,5 @@ import { useMulticallContract } from './useContract'
 // gets the current timestamp from the blockchain
 export default function useCurrentBlockTimestamp(): bigint | undefined {
   const multicall = useMulticallContract()
-  return useSingleCallResult(multicall, 'getCurrentBlockTimestamp')?.result as bigint
+  return useSingleCallResult(multicall, 'getCurrentBlockTimestamp')?.result
 }
